backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
ebooklib
beautifulsoup4
python-docx
httpx
python-dotenv
